backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
python-jose==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
openai==1.100.2
python-multipart==0.0.9
